Store all houses in a component state
This commit gives state an initial value by passing that value to the
function which in this case is an empty array. This uses a form of
destructuring taking the value and the function out of the array.
Finally, the commit calls setAllHouses when they are fetched which
stores all houses in a component state