Create a new route component
This commit first installs react-router-dom and import its relevant
components
This commit gets rid of the existing JSX in the return statement of
the component and creating a new route component using the imported
router component. Then directly within the router, JSX is created
After this, the commit uses the Switch component from the react router
from which underneath it, the Route component is used to specify the
route path. Switch checks the child route component to see if there's
a match between the actual URL in the browser and the path specified
in the route.